+++ truffle/src/com.oracle.truffle.api.instrumentation/src/com/oracle/truffle/api/instrumentation/TruffleInstrument.java
@@ -147,6 +147,9 @@ public abstract class TruffleInstrument {
      * onCreate} method:
      *
      * {@codesnippet DebuggerExample}
+     * <p>
+     * If this method throws an {@link com.oracle.truffle.api.exception.AbstractTruffleException}
+     * the exception interop messages are executed without a context being entered.
      *
      * @param env environment information for the instrument
      *
@@ -403,7 +406,9 @@ public abstract class TruffleInstrument {
          * Returns a new value for a context local of an instrument. The returned value must not be
          * <code>null</code> and must return a stable and exact type per registered instrument. A
          * thread local must always return the same {@link Object#getClass() class}, even for
-         * multiple instances of the same {@link TruffleInstrument}.
+         * multiple instances of the same {@link TruffleInstrument}. If this method throws an
+         * {@link com.oracle.truffle.api.exception.AbstractTruffleException} the exception interop
+         * messages may be executed without a context being entered.
          *
          * @see TruffleInstrument#createContextLocal(ContextLocalFactory)
          * @since 20.3
@@ -424,7 +429,9 @@ public abstract class TruffleInstrument {
          * value must not be <code>null</code> and must return a stable and exact type per
          * registered instrument. A thread local must always return the same
          * {@link Object#getClass() class}, even for multiple instances of the same
-         * {@link TruffleInstrument}.
+         * {@link TruffleInstrument}. If this method throws an
+         * {@link com.oracle.truffle.api.exception.AbstractTruffleException} the exception interop
+         * messages may be executed without a context being entered.
          *
          * @see TruffleInstrument#createContextThreadLocal(ContextThreadLocalFactory)
          * @since 20.3